Moved NavBar to Widget
