+++ apps/verification-portal/src/lib/widgets/Navbar/HamburgerButton.svelte
