+++ apps/verification-portal/src/lib/widgets/ThemeSwitch/Menu.svelte
@@ -3,6 +3,7 @@
   import LL from '$lib/shared/i18n/i18n-svelte';
   import RolloverBtn from '$lib/shared/components/RolloverBtn.svelte';
   import { themeStore, setTheme } from '$lib/features/themeSwitch';
+  import type { Theme } from '$lib/features/themeSwitch';
   export const btnClass: string =
     ' text-gray-500 dark:text-gray-400  hover:bg-gray-100 dark:hover:bg-gray-700 focus:outline-none focus:ring-4 focus:ring-gray-200 dark:focus:ring-gray-700  cursor-pointer ';
 
@@ -17,7 +18,7 @@
 
   let dropdownOpen = false;
   const handleThemeChange = () => {
-    setTheme(selectedTheme);
+    setTheme(selectedTheme as Theme);
 
     dropdownOpen = false;
   };
