+++ apps/verification-portal/src/routes/+layout.svelte
@@ -4,7 +4,7 @@
   import { fade } from 'svelte/transition';
   import { onMount } from 'svelte';
   import { Toaster } from '@sni/ui-kit';
-  import NavBar from '$lib/components/navbar/index.svelte';
+  import NavBar from '$lib/widgets/Navbar/NavBar.svelte';
   import Footer from '$lib/components/Footer.svelte';
   import { afterNavigate, beforeNavigate } from '$app/navigation';
   import {
